Update: Improve reporting and structure of RateControl and related modules Add: testing for report generation Add: warning for windows async
