+++ fspin/__init__.py
@@ -1,3 +1,3 @@
-from .RateControl import RateControl as rate
-from .RateControl import spin
-from .RateControl import loop
+from .rate_control import RateControl as rate
+from .decorators import spin
+from .loop_context import loop
